+++ Plugins/RHI/PS2/src/Ps2Texture.cpp
@@ -224,8 +224,8 @@ void Ps2Texture::Bind() const {
     lod_t lod{};
     lod.calculation = LOD_USE_K;
     lod.max_level = 0;
-    lod.mag_filter = LOD_MAG_NEAREST;
-    lod.min_filter = LOD_MIN_NEAREST;
+    lod.mag_filter = LOD_MAG_LINEAR;
+    lod.min_filter = LOD_MIN_LINEAR;
     lod.l = 0;
     lod.k = 0;