+++ Projects/Ps2ThirdPerson/Ps2ThirdPersonDemo.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace leon {
+class Window;
+}
+
+namespace leon::ps2thirdperson {
+
+[[nodiscard]] int RunPs2ThirdPersonDemo(Window& window);
+
+} // namespace leon::ps2thirdperson